refactor: firebase service and angular-fire removal
