+++ package.json
@@ -36,7 +36,6 @@
     "@angular/common": "~17.2.2",
     "@angular/core": "~17.2.2",
     "@angular/elements": "^17.2.2",
-    "@angular/fire": "17.0.1",
     "@angular/forms": "~17.2.2",
     "@angular/platform-browser": "~17.2.2",
     "@angular/platform-browser-dynamic": "~17.2.2",
